Create Money.ps1
This script works as a module to allow managing the personal coins one carries, what you may have in a bank you're not carrying with you, and a party fund.  If the DM also charges tax, you can put that into here too, and keep track of that.  It writes each transaction to a log, and can work with multiple profiles/PCs/Parties.